+++ src/Helpers/generate-layout.types.ts
@@ -3,7 +3,20 @@ import type { ResponsiveStyleValue } from '@mui/system'
 import type { Theme } from '@mui/material'
 
 export type LayoutPositions = 'top-left' | 'top-center' | 'top-right' | 'center-left' | 'center-right' | 'bottom-left' | 'bottom-center' | 'bottom-right' | 'inherit'
-export type GenerateLayoutPositionsProps = { propValue: LayoutPositions }
+
+export type GenerateLayoutPositionsProps = {
+	/**
+	 * Position of the component
+	 */
+	propValue: LayoutPositions
+
+	/**
+	 * Determines whether there should be a margin to the right of the button component
+	 *
+	 * @default false
+	 */
+	rightHandSide: boolean
+}
 
 export type GenerateLayoutProps = {
 	/**
@@ -13,7 +26,15 @@ export type GenerateLayoutProps = {
 
 	/**
 	 * Defines the positioning of the button component
+	 *
 	 * @default undefined
 	 */
-	layout?: ResponsiveStyleValue<LayoutPositions> | undefined
+	layout: ResponsiveStyleValue<LayoutPositions> | undefined
+
+	/**
+	 * Determines whether there should be a margin to the right of the button component
+	 *
+	 * @default false
+	 */
+	rightHandSide: boolean
 }